requirements.txt
twine==3.2.0
nltk==3.5
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
DateTime==4.3
diffbot==2.0.0
pandas==1.1.1
numpy==1.19.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
Flask==1.1.2
gensim==3.8.3
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
tweepy==3.9.0
googletrans==3.0.0
scikit-learn==0.23.2
plotly
matplotlib==3.3.1
argparse
pytest==6.0.1
